doc: upload and delete file
