+++ doc/docs.go
@@ -6,10 +6,10 @@ import "github.com/swaggo/swag/v2"
 
 const docTemplate = `{
     "schemes": {{ marshal .Schemes }},
-    "components": {"schemas":{"dto.ChatDirectResponse":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/dto.DirectMessageResponse"},"type":"array","uniqueItems":false},"user":{"$ref":"#/components/schemas/dto.PublicUserResponse"}},"type":"object"},"dto.DirectMessageResponse":{"properties":{"content":{"type":"string"},"id":{"type":"integer"},"receiverId":{"type":"integer"},"sendedAt":{"type":"string"},"senderId":{"type":"integer"}},"type":"object"},"dto.HttpError":{"properties":{"error":{"type":"string"}},"type":"object"},"dto.HttpListResponse-dto_ChatDirectResponse":{"properties":{"result":{"items":{"$ref":"#/components/schemas/dto.ChatDirectResponse"},"type":"array","uniqueItems":false}},"type":"object"},"dto.HttpResponse-dto_LoginResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.LoginResponse"}},"type":"object"},"dto.HttpResponse-dto_RegisterResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.RegisterResponse"}},"type":"object"},"dto.HttpResponse-dto_TokenResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.TokenResponse"}},"type":"object"},"dto.HttpResponse-dto_UserResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.LoginRequest":{"properties":{"idToken":{"type":"string"},"provider":{"description":"GOOGLE","type":"string"}},"required":["idToken","provider"],"type":"object"},"dto.LoginResponse":{"properties":{"accessToken":{"type":"string"},"exp":{"type":"integer"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.PublicUserResponse":{"properties":{"email":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"profilePictureUrl":{"type":"string"}},"type":"object"},"dto.RefreshTokenRequest":{"properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"],"type":"object"},"dto.RegisterRequest":{"properties":{"idToken":{"type":"string"},"provider":{"description":"GOOGLE","type":"string"}},"required":["idToken","provider"],"type":"object"},"dto.RegisterResponse":{"properties":{"accessToken":{"type":"string"},"exp":{"type":"integer"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.TokenResponse":{"properties":{"accessToken":{"type":"string"},"exp":{"type":"integer"},"refreshToken":{"type":"string"}},"type":"object"},"dto.UserResponse":{"properties":{"email":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"profilePictureUrl":{"type":"string"}},"type":"object"}},"securitySchemes":{"ApiKeyAuth":{"in":"header","name":"Authorization","type":"apiKey"}}},
+    "components": {"schemas":{"dto.ChatDirectResponse":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/dto.DirectMessageResponse"},"type":"array","uniqueItems":false},"user":{"$ref":"#/components/schemas/dto.PublicUserResponse"}},"type":"object"},"dto.DirectMessageResponse":{"properties":{"content":{"type":"string"},"id":{"type":"integer"},"receiverId":{"type":"integer"},"sendedAt":{"type":"string"},"senderId":{"type":"integer"}},"type":"object"},"dto.FileDetailResponse":{"properties":{"id":{"type":"integer"},"ownerId":{"type":"integer"},"url":{"type":"string"}},"type":"object"},"dto.HttpError":{"properties":{"error":{"type":"string"}},"type":"object"},"dto.HttpListResponse-dto_ChatDirectResponse":{"properties":{"result":{"items":{"$ref":"#/components/schemas/dto.ChatDirectResponse"},"type":"array","uniqueItems":false}},"type":"object"},"dto.HttpResponse-dto_FileDetailResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.FileDetailResponse"}},"type":"object"},"dto.HttpResponse-dto_LoginResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.LoginResponse"}},"type":"object"},"dto.HttpResponse-dto_TokenResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.TokenResponse"}},"type":"object"},"dto.HttpResponse-dto_UserResponse":{"properties":{"result":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.LoginRequest":{"properties":{"idToken":{"type":"string"},"provider":{"description":"GOOGLE","type":"string"}},"required":["idToken","provider"],"type":"object"},"dto.LoginResponse":{"properties":{"accessToken":{"type":"string"},"exp":{"type":"integer"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.PublicUserResponse":{"properties":{"email":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"profilePictureUrl":{"type":"string"}},"type":"object"},"dto.RefreshTokenRequest":{"properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"],"type":"object"},"dto.TokenResponse":{"properties":{"accessToken":{"type":"string"},"exp":{"type":"integer"},"refreshToken":{"type":"string"}},"type":"object"},"dto.UserResponse":{"properties":{"email":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"profilePictureUrl":{"type":"string"}},"type":"object"},"dto.UserUpdateRequest":{"properties":{"name":{"type":"string"}},"type":"object"}},"securitySchemes":{"ApiKeyAuth":{"in":"header","name":"Authorization","type":"apiKey"}}},
     "info": {"description":"{{escape .Description}}","title":"{{.Title}}","version":"{{.Version}}"},
     "externalDocs": {"description":"","url":""},
-    "paths": {"/api/v1/auth/login":{"post":{"description":"Login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.LoginRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_LoginResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"summary":"Login","tags":["auth"]}},"/api/v1/auth/logout":{"post":{"description":"Logout","responses":{"204":{"description":"No Content"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Logout","tags":["auth"]}},"/api/v1/auth/refresh-token":{"post":{"description":"Refresh Token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.RefreshTokenRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_TokenResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"summary":"Refresh Token","tags":["auth"]}},"/api/v1/auth/register":{"post":{"description":"Register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.RegisterRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_RegisterResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"summary":"Register","tags":["auth"]}},"/api/v1/me":{"get":{"description":"Get me","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_UserResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Get me","tags":["user"]}},"/api/v1/messages":{"get":{"description":"list all direct message chats for each individual user","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpListResponse-dto_ChatDirectResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"list all direct message chats","tags":["direct_message"]}},"/api/v1/messages/ws":{"get":{"description":"Establish a WebSocket connection for real-time communication.\nThe message from a server will be in a format of \"[EVENT] [MESSAGE]\" which [EVENT] can be \"ERROR\" or \"MESSAGE\".\nIf [EVENT] is error, [MESSAGE] will be a string of error message ,otherwise it will be a dto.RealTimeMessageResponse","responses":{"101":{"description":"Switching Protocols"},"400":{"description":"Bad Request"}},"security":[{"ApiKeyAuth":[]}],"summary":"connect to websocket","tags":["message"]}}},
+    "paths": {"/api/v1/auth/login":{"post":{"description":"Login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.LoginRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_LoginResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"summary":"Login","tags":["auth"]}},"/api/v1/auth/logout":{"post":{"description":"Logout","responses":{"204":{"description":"No Content"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Logout","tags":["auth"]}},"/api/v1/auth/refresh-token":{"post":{"description":"Refresh Token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.RefreshTokenRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_TokenResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"summary":"Refresh Token","tags":["auth"]}},"/api/v1/files":{"post":{"requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"type":"file"}}},"description":"file","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_FileDetailResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"upload file","tags":["files"]}},"/api/v1/files/{fileID}":{"delete":{"parameters":[{"description":"file id","in":"path","name":"fileID","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"delete file","tags":["files"]}},"/api/v1/me":{"get":{"description":"Get me","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_UserResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Get me","tags":["user"]},"patch":{"description":"Update user's profile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.UserUpdateRequest"}}},"description":"request request","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpResponse-dto_UserResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"Update me","tags":["user"]}},"/api/v1/messages":{"get":{"description":"list all direct message chats for each individual user","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpListResponse-dto_ChatDirectResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HttpError"}}},"description":"Internal Server Error"}},"security":[{"ApiKeyAuth":[]}],"summary":"list all direct message chats","tags":["direct_message"]}},"/api/v1/messages/ws":{"get":{"description":"Establish a WebSocket connection for real-time communication.\nThe message from a server will be in a format of \"[EVENT] [MESSAGE]\" which [EVENT] can be \"ERROR\" or \"MESSAGE\".\nIf [EVENT] is error, [MESSAGE] will be a string of error message ,otherwise it will be a dto.RealTimeMessageResponse","responses":{"101":{"description":"Switching Protocols"},"400":{"description":"Bad Request"}},"security":[{"ApiKeyAuth":[]}],"summary":"connect to websocket","tags":["message"]}}},
     "openapi": "3.1.0"
 }`
