+++ doc/swagger.yaml
@@ -23,6 +23,15 @@ components:
         senderId:
           type: integer
       type: object
+    dto.FileDetailResponse:
+      properties:
+        id:
+          type: integer
+        ownerId:
+          type: integer
+        url:
+          type: string
+      type: object
     dto.HttpError:
       properties:
         error:
@@ -36,15 +45,15 @@ components:
           type: array
           uniqueItems: false
       type: object
+    dto.HttpResponse-dto_FileDetailResponse:
+      properties:
+        result:
+          $ref: '#/components/schemas/dto.FileDetailResponse'
+      type: object
     dto.HttpResponse-dto_LoginResponse:
       properties:
         result:
           $ref: '#/components/schemas/dto.LoginResponse'
-      type: object
-    dto.HttpResponse-dto_RegisterResponse:
-      properties:
-        result:
-          $ref: '#/components/schemas/dto.RegisterResponse'
       type: object
     dto.HttpResponse-dto_TokenResponse:
       properties:
@@ -96,28 +105,6 @@ components:
       required:
       - refreshToken
       type: object
-    dto.RegisterRequest:
-      properties:
-        idToken:
-          type: string
-        provider:
-          description: GOOGLE
-          type: string
-      required:
-      - idToken
-      - provider
-      type: object
-    dto.RegisterResponse:
-      properties:
-        accessToken:
-          type: string
-        exp:
-          type: integer
-        refreshToken:
-          type: string
-        user:
-          $ref: '#/components/schemas/dto.UserResponse'
-      type: object
     dto.TokenResponse:
       properties:
         accessToken:
@@ -136,6 +123,11 @@ components:
         name:
           type: string
         profilePictureUrl:
+          type: string
+      type: object
+    dto.UserUpdateRequest:
+      properties:
+        name:
           type: string
       type: object
   securitySchemes:
@@ -239,22 +231,21 @@ paths:
       summary: Refresh Token
       tags:
       - auth
-  /api/v1/auth/register:
+  /api/v1/files:
     post:
-      description: Register
       requestBody:
         content:
-          application/json:
+          application/x-www-form-urlencoded:
             schema:
-              $ref: '#/components/schemas/dto.RegisterRequest'
-        description: request request
+              type: file
+        description: file
         required: true
       responses:
         "200":
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/dto.HttpResponse-dto_RegisterResponse'
+                $ref: '#/components/schemas/dto.HttpResponse-dto_FileDetailResponse'
           description: OK
         "400":
           content:
@@ -262,15 +253,58 @@ paths:
               schema:
                 $ref: '#/components/schemas/dto.HttpError'
           description: Bad Request
+        "401":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Unauthorized
         "500":
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/dto.HttpError'
           description: Internal Server Error
-      summary: Register
+      security:
+      - ApiKeyAuth: []
+      summary: upload file
       tags:
-      - auth
+      - files
+  /api/v1/files/{fileID}:
+    delete:
+      parameters:
+      - description: file id
+        in: path
+        name: fileID
+        required: true
+        schema:
+          type: integer
+      responses:
+        "200":
+          description: OK
+        "401":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Unauthorized
+        "404":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Not Found
+        "500":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Internal Server Error
+      security:
+      - ApiKeyAuth: []
+      summary: delete file
+      tags:
+      - files
   /api/v1/me:
     get:
       description: Get me
@@ -296,6 +330,39 @@ paths:
       security:
       - ApiKeyAuth: []
       summary: Get me
+      tags:
+      - user
+    patch:
+      description: Update user's profile
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/dto.UserUpdateRequest'
+        description: request request
+        required: true
+      responses:
+        "200":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpResponse-dto_UserResponse'
+          description: OK
+        "400":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Bad Request
+        "500":
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/dto.HttpError'
+          description: Internal Server Error
+      security:
+      - ApiKeyAuth: []
+      summary: Update me
       tags:
       - user
   /api/v1/messages:
