+++ internal/dto/file.go
@@ -1,0 +1,7 @@
+package dto
+
+type FileDetailResponse struct {
+	ID      uint   `json:"id"`
+	URL     string `json:"url"`
+	OwnerID uint   `json:"ownerId"`
+}
