+++ internal/server/register_routes.go
@@ -19,7 +19,6 @@ func (s *Server) RegisterRoutes(
 	// auth
 	auth := v1.Group("/auth")
 	auth.Post("/login", authHandler.HandleLogin)
-	auth.Post("/register", authHandler.HandleRegister)
 	auth.Post("/refresh-token", authHandler.HandleRefreshToken)
 	auth.Post("/logout", authMiddleware.Auth, authHandler.HandleLogout)
